Create ExtraInfoDisplay. Add action to InputMap.
This commit creates ExtraInfoDisplay and all associated resources.
It creates a node in HUD for keeping nodes hidden when HUD's show() is
called. Adds ExtraInfoDisplay.tscn to HUD. The action
hide_extra_info_display has been added to the InputMap via project
settings and has the keyboard key 4 as it's InputEventKey.